+++ src/src/io.c
@@ -82,6 +82,7 @@ void initialize_person(struct person *pnew) {
   pnew->dmult = 1.0;
   pnew->tmult = 1.0;
   pnew->down = NULL;
+  pnew->marriage_queue_index = -1;
   pnew->extra = NULL;
   pnew->ltrans = NULL;
 }
@@ -555,6 +556,7 @@ void fix_pop_pointers()
     }
 
     p->pointer_type[MARRIAGE_QUEUE] = PTR_NULL;
+    p->marriage_queue_index = -1;
     p->pref = random_spouse2;
     p->score = score3;
 #ifdef ENHANCED